+++ src/pages/WarehouseReceipt.jsx
